все скрипты на питоне были разбиты на папки, догружен датасет
