+++ python_vizualizations/detailed_remainings.py
